删除用户功能更新

+++ omgcms/src/main/java/com/omgcms/repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import com.omgcms.model.core.User;
 
-public interface UserRepository extends JpaSpecificationExecutor<User>, JpaRepository<User, Integer> {
+public interface UserRepository extends JpaSpecificationExecutor<User>, JpaRepository<User, Long> {
 
 	// 根据 userAccount 来获取对应的 User
 	public User getByUserAccount(String userAccount);
